requirements.txt
streamlit
pymongo
pandas
numpy
Pillow
